RoaderBot-api/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API for RoaderBot"
authors = ["komediantto <komediantto@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
